refactor(gui): stage config utils
Extract utils for converting raw stage configs to new file
